Refactor imports for dashboard components
