+++ client/src/components/home/home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MainNavbar from '../navigation/main-navbar.jsx';
+import MainNavbar from '../navigation/MainNavbar';
 
 class Home extends React.Component {
   constructor(props) {
